requirements.txt
flask==2.3.2
python-telegram-bot==20.3
requests==2.31.0